Exchange parameter order of Option<T>.map/mapOrElse.
These APIs return the result of apply a given function to inner value
or the one of default value.

These names are the order of 'Map' 'Default Operations'.
So use might be expect that user should pass 'map function' and 'default' in such order.

However, the current APIs' signature don't follow the rule
because original Rust's APIs use 'default' and 'mapFn' order.

I think this is bad order in JavaScript worlds, this world has
condirional operator '?:', it is 'true ? "true case" : "false" case'.
Therefore it should be the expected parameter order.