Add OptionT.OptionBase & Mark OptionT.OptionT as a deprecated API.
This library considers 2 type systems, ECMA262 5th' one and TypeScript's one.
So there are some way to detect whether this object is a `OptionT.Option<T>`:

In ECMA262 5th, there is no interface system. Thus we can detect the
object type by checking `__proto__` like this.

```
var isOptionT = some instanceof OptionT.OptionT.
```

In TypeScript, there is interface system and static type checking!
So we can use interface to do it and it is checked statically.

```
function isSome<T>(option: Option<T>): boolean {.......}
```

By the way, this might be a problem to confuse "what API should use in
our script? `OptionT.OptionT`? or `OptionT.Option<T>?`.

So this change makes the name of 'base class' more explicitly
and marks the legacy name as deprecated.