Expose OptionBase in the type definition.
Generally, in some static typing environments as TypeScript,
Flowtype, and etc, we should use an interface strongly,
not to use class base typing  like `instanceof`.

However, if we cooperate with some library that provides
the runtime assertion which uses `a instanceof A` like React's `React.PropTypes`
for cooperating with some other vanilla JavaScript ones,
  we requires the checking systems which is based on `instanceof`.

So this commits exposes `OptionBase` to our type definition to
cooperate with them. This is needed for real world usecases.

The typical case is TSX (TypeScript JSX) syntax.
https://github.com/Microsoft/TypeScript/wiki/JSX