Add the comment why we check the returned value type from selector in map operation
If `U` is `Maybe<SomeType>`, we think naturally the returned value of this function would be
the nested type `Maybe<Maybe<SomeType>>`. But this type means `(SomeType | null | undefined) | null | undefined`.
So a type checker would recognize this type as `SomeType | null | undefined`. So it's flattened.
Then the user should call `andThen` (_flatmap_) operation instead of this.

And we should throw the error for it.