Exports all existing pathes by `exports` field to make the compatiblity.
By the implementation of Node.js v13.8.0, if the package.json has
`exports` even if it is an empty object, we don't import modules by a
subpath.

This means we cannot use `require('option-t/cjs/Nullable')`
or `import {} from 'option-t/esm/Nullable/index.mjs'`. This causes breaking change.

To avoid the breaking change, we need to all existing paths to `exports`.