Fix `ERR_PACKAGE_PATH_NOT_EXPORTED` error on loading `option-t`.
This error happens by this change in [Node,js v13.10](https://github.com/nodejs/node/pull/31625) and I think we can regard this problem as the regression of this library.

For the future, we should wait https://github.com/nodejs/node/issues/32107 but it is still in progress.
This patch will try to workaround for it.

* https://github.com/nodejs/node/pull/31625
* https://github.com/nodejs/node/issues/32107
* https://github.com/babel/babel/issues/11216