Drop CircleCI settings
I still think CircleCI is a one of the best CI solution
if we develops commercial products. It is easy setting fork/join for jobs,
powerful (aweful) yaml, and well-cached docker images, and more.
They works well with nice throughput if we pay to their service.

However, for non-commercial open source project, I don't think CircleCI
is best choise now (at leeast for us). We need sign-in to see the CI
result and some management console is behind their auth gateway....

And we have a GitHub Actions now. This has some pros compared with
CircleCI. But this is well integrated with GitHub.

For closed source repositories hosted on GitHub, GitHub Actions still
lacks some foundamental key points to work as distributed VCS which each
team member can have their repositories.

But option-t is open source project. We can avoid this foundamental
problem.

So it's the time to switch.