Mark `ok`/`err` export from PlainResult/index as deprecated.
We would like to remove these.
In Rust, `Option<T>` is a 1st citizen. So I think it's beneficial to
provide conversions as a short name.

But this is not.
Generally, `null` may be more better choice in JavaScript.
So the time is over to treat it as special case.