Replace cpx with the simple node.js script
We'd like we'd like to clear GitHub's noisey vulnerable alert which says that
cpx's dependency contains some vulnerable packages.
(This alert does not affect this package actually)

But It looks that [cpx](https://github.com/mysticatea/cpx) has not been updated since 2018.

So we replace it with simple node.js script