Mark PlainOption as "weak" deprecated
In JavaScript world, simple nullable type is better
to express an absence of a value than this kind of tagged union type.

`T | null` or `T | undefined` are common idiom for almost use-cases, are interoperable,
and make a thing simple. Furthermore we provide a utility toolkit for their types.

This kind of type is still useful to contains null or undefined as not an absence value.
But it’s a rare case. There is no reason to recommend to use it in general use cases.

Thus, now, we stop to recommend to use this and mark it as deprecated.

This does not mean that we remove this type in near future.
It’s just to draw down from a recommended types.

Probably we will not be able to remove this for backward compatibility.
Furthermore, it would not be costly to keep this type because:

1. This type does not affect a bundle size by tree shaking support.
2. The implementation does not make our build pipeline complex.