Implement alternative that does not use `PlainOption` for `ClassicOption/compat`
We start to mark `PlainOption` as deprecated since 23bbaa1c426f4a978122cf32a8a3510f49098d6b.
We don't have any plan to remove it (and I think we will not it),
but I think we should provide an alterntive operator that
does not use `PlainOption`. For example, `ClassicOption/compat`.