Set `Error.cause` in `assertIsErrorInstance()` by hand if it is not implemented by the host environment
I prefer this solution than #1831.
Fix #1832

Motivation
---------------

Why we call `String(input)` in `assertIsErrorInstance()`
https://github.com/option-t/option-t/blame/18b119a2bab1bda6c68e845a544efa523d243994/src/internal/assert.ts#L15-L16

is to improve debuggability on devtools.
But I slacked to implement it with some compat path with `Error.cause`.
As the result, we need to create a patch as #1831.

This patch rewrite the path for non `Error` instance in `assertIsErrorInstance()`
with implementing `Error.cause` polyfill (Of cource we would like to remove it in the future).

This simplify the implementation and take more standard like way.

Drawbacks
----------

Performance
============

This path is a fallback path for the place which we expect coming an `Error` instance.
I think we can allow some slow down.

Changing an error message format
=============

This is a pain point but I think this would not be marked
as a breaking change because the almost behavior is not changed.
However, this would be a problem if user's observability tool
does not support `Error.cause`. But it's a another problem, I think.