Expose `unwrapOrThrowError` from `plain_result` or `plain_result/namespace`.
We would not like to expose them due to recommend to handle result type in that style.

However, react 19 introduce _Action_ as their terminology as a part of thier framework for async transition.
By this change, some of their APIs allow to take an async function that throwing an exception.
https://19.react.dev/blog/2024/04/25/react-19#actions

To improve an ergonomics with them, we expose `unwrapOrThrowError()` as tier 1 operator.