Add variants of tryCatchIntoResult that wrapping automatically a thrown something with current realm's Error if it is not an Error instance of current realm
This adds these new APIs as experimental:

- `tryCatchIntoResultAndMapErrFallback` from `option-t/plain_result/experimental/try_catch`.
- `tryCatchIntoResultAndMapErrFallbackAsync` from `option-t/plain_result/experimental/try_catch_async`

This experiments to try to resolve a solution for #2286.

Motivation
-----------------

The basic motivations is same with ones written in #2286.

Typical case to generate `Result<T, unknown>` is originated from enclosure
a exist throw-exception style codebase to bridge it to _Result_.
I think we should care only its case. The others can be solved by `mapErr`.

Design
--------------

If _producer_ throws something, the result will be `Err(unknown)`.
Then, we check it by `value instanceof Error`.

- If _the thrown value_ is an `Error` instance of the current [realm][realm],
  return it directly and the result will be `Err(Error)`.
    - The check happens in the same operator with calling _producer_.
      I think we don't have to wrap with an extra error always.
- If _the thrown value_ is **not** an `Error` instance of the current [realm][realm] (including `number`, `string`, `boolean` or others),
  we wrap it by `UnknownCausalError` introduced by this, and the result will be `Err(UnknownCausalError)`.
    - But we don't expose this `UnknownCausalError` type to the type signature.

[realm]: https://262.ecma-international.org/14.0/#realm